Lots of updates to Tetris game Sequence System to engine Rewrote animation sprite sheet format to parse  asesprite sheets

+++ src/Engine/Core/Animation.cpp
@@ -1,6 +1,6 @@
 #include "./Animation.h"
 
-Animation::Animation(std::string name, int index, int totalFrames, int speed) {
+Animation::Animation(std::string name, int index, int totalFrames, float speed) {
     this->name = name;
     this->index = index;
     this->totalFrames = totalFrames;
